Here's everything from 17 y.o Victor
I don't want to touch the code, since it's like a snapshot in time. I
found this on a hard-drive and genuinely thought I lost this code. But
it looks like it's all here :D.

I didn't know git at the time I was creating Bloopy, so this had to be a
brand new repository. Fun fact, this was my second attempt at creating
Bloopy, since my first attempt was lost in a hard drive failure. It
would've helped if I knew git then huh? It all worked out though, and,
tbh, this was much better than my first attempt. It's not up to my
standards nowadays, but I wouldn't be where I am today without this app
:)

It's not on Google Play anymore, but if you're up for it, I'm sure you
can generate an APK, granted, this was built on API version 14 or some
nonsense like that. So, good luck :D